+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Added:
 
 - LUA script `float.all.samesizze.lua`script 
+  - `SUPER + CTRL + SPACE` to activate
   - sets the sizes based on number of windows and resolution
 - Documented `fastfetch` `config.json` on how to add graphical logo
 - `ToggleOpactiy.sh`